Fix for issue #215
This fixes the issue with the test page--I went through all the other
uses of curl in the project, and usually it makes calls to external
APIs, not local things. I wasn't sure what the use case for the curl
call in Utilities.php was, but I added a user agent there as well in
case it ever gets used to call anything local.
Also, I closed the curl session twice in test.php--looks like that was
left out (although it's not a big deal).

Wasn't sure what to use for a user agent, so I chose an old standard.
Any user agent would have fixed the issue, however.