Stripe Currency
Add option to set stripe currency from env and have this pass through the app into Stripe to allow the button structure to appear correct.

The default currency is USD, if no currency is passed the [ PAY ] button indicates USD even if in the stripe dashboard it is a different amount. By passing through the currency this removes the USD from the [ PAY ] button not confusing the customer